requirements.txt
slack_bolt
python-dotenv
watchfiles
openai
requests